UPD: merged back EXTERNAL_TRANSCODES branch
Main changes:
  All drivers are now Plugins
  Plugins are loaded once a driver is requested
  search location is either ini-entry DRIVERPLUGINDIRS, envvar VISTACORELIBS_DRIVER_PLUGIN_DIRS, or .
RENAME: all base types (uint32, microtime, ...) are now in VistaType namespace

git-svn-id: https://svn.rwth-aachen.de/repos/vrgroup-svn/projects/Vista/trunk@22128 33c38b33-09d2-4561-8d4d-1bfa2f58f885

+++ VistaCoreLibs/VistaInterProcComm/IPNet/VistaSocket.cpp
@@ -929,8 +929,8 @@ unsigned long IVistaSocket::WaitForIncomingData(int iTimeout)
 	}
 	if(iRet == -1)
 	{
-		printf("-- WARNING -- socket-select returned -1 (exception) during WaitForIncomingData(%d, %d).\n",
-				m_iSocketID, iTimeout);
+		printf("-- WARNING -- socket-select returned -1 (exception) during WaitForIncomingData(%ld, %d).\n",
+				long(m_iSocketID), iTimeout);
 #if defined(VISTA_IPC_USE_EXCEPTIONS)
 		VISTA_THROW("IVistaSocket::WaitForIncomingData -- Exception", 0x00000104)
 #endif
